+++ wden/src/bitwarden/mod.rs
@@ -1,3 +1,4 @@
 pub mod api;
 pub mod cipher;
+pub mod keys;
 pub mod server;